Revert "Restructure templates into marketing and app areas"
