+++ src/main/java/com/project/tracking_system/configuration/SecurityConfiguration.java
@@ -83,11 +83,10 @@ public class SecurityConfiguration {
                                 .policy(ReferrerPolicyHeaderWriter.ReferrerPolicy.SAME_ORIGIN))
                 )
                 .authorizeHttpRequests(authorizeRequests -> authorizeRequests
-                        .requestMatchers("/", "/features", "/pricing", "/login", "/logout", "/registration", "/forgot-password", "/reset-password",
+                        .requestMatchers("/", "/login", "/logout", "/registration", "/forgot-password", "/reset-password",
                                 "/privacy-policy", "/terms-of-use", "/css/**", "/js/**", "/bootstrap/**", "/images/**",
                                 "/upload", "/ws/**", "/wss/**", "/sample/**", "/download-sample", "/tariffs", "/tariffs/**").permitAll()
                         .requestMatchers("/admin/**").hasRole("ADMIN")
-                        .requestMatchers("/app/**").authenticated()
                         .anyRequest().authenticated()
                 )
                 .formLogin(formLogin -> formLogin
@@ -108,7 +107,7 @@ public class SecurityConfiguration {
                                     return;
                                 }
                             }
-                            response.sendRedirect("/app");
+                            response.sendRedirect("/");
                         })
                         .failureHandler((request, response, exception) -> {
                             String email = request.getParameter("email");
