+++ src/main/java/com/project/tracking_system/controller/AnalyticsController.java
@@ -149,7 +149,7 @@ public class AnalyticsController {
 
         model.addAttribute("nonce", request.getAttribute("nonce"));
 
-        return "app/analytics/dashboard";
+        return "analytics/dashboard";
     }
 
     /**
