+++ src/main/java/com/project/tracking_system/controller/HomeController.java
@@ -26,8 +26,8 @@ import java.util.List;
 @RequiredArgsConstructor
 @Slf4j
 @Controller
-@RequestMapping("/app")
-public class AppController {
+@RequestMapping("/")
+public class HomeController {
 
     private final TrackFacade trackFacade;
     private final StoreService storeService;
@@ -44,7 +44,7 @@ public class AppController {
             List<Store> stores = storeService.getUserStores(user.getId());
             model.addAttribute("stores", stores);
         }
-        return "app/home";
+        return "home";
     }
 
     /**
@@ -84,7 +84,7 @@ public class AppController {
             if (trackInfo == null || trackInfo.getList().isEmpty()) {
                 model.addAttribute("customError", "Нет данных для указанного номера посылки.");
                 log.warn("Нет данных для номера: {}", number);
-                return "app/home";
+                return "home";
             }
 
             model.addAttribute("trackInfo", trackInfo);
@@ -95,7 +95,19 @@ public class AppController {
             model.addAttribute("generalError", "Произошла ошибка при обработке запроса.");
             log.error("Общая ошибка: {}", e.getMessage(), e);
         }
-        return "app/home";
+
+        return "home";
+    }
+
+
+    @GetMapping("/privacy-policy")
+    public String privacyPolicy() {
+        return "privacy-policy";
+    }
+
+    @GetMapping("/terms-of-use")
+    public String termsOfUse() {
+        return "terms-of-use";
     }
 
 }
