+++ src/main/resources/templates/analytics/dashboard.html
