+++ src/main/resources/templates/departures.html
