--- src/main/resources/templates/marketing/features.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/layout}">
-<head>
-    <title layout:fragment="title">Возможности</title>
-</head>
-<body>
-<main layout:fragment="content">
-    <div class="container">
-        <h1>Возможности</h1>
-        <p>Здесь описаны преимущества вашего приложения.</p>
-    </div>
-</main>
-</body>
-</html>
