--- src/main/resources/templates/marketing/pricing.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/layout}">
-<head>
-    <title layout:fragment="title">Цены</title>
-</head>
-<body>
-<main layout:fragment="content">
-    <div class="container">
-        <h1>Тарифы</h1>
-        <p>Информация о ценах и подписках.</p>
-    </div>
-</main>
-</body>
-</html>
