+++ src/main/resources/templates/partials/cookies.html
