+++ src/main/resources/templates/partials/customer-info.html
