+++ src/main/resources/templates/partials/footer.html
