+++ src/main/resources/templates/partials/store-statistics.html
