+++ src/main/resources/templates/terms-of-use.html
